pyproject.toml
[project]
name = "barbershop"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex",email = "alex70.07@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "markdown (>=3.7,<4.0)",
    "openai (>=1.70.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=5.1.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
